remove nextui and unused components
