+++ apps/devprofile/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@czstr/ui": "*",
-    "@nextui-org/react": "^1.0.8-beta.5",
     "next": "^13.0.0",
     "node-fetch": "^2.6.1",
     "react": "^18.2.0",
