+++ apps/devprofile/public/assets/turing-logo.png
